Fix browser package scanning
PackageManager.queryIntentActivities is documented as always returning
matching activities in a priority order, such that the default match
would always be returned first. Therefore, we do not need to
separately handle default browsers: scanning in order and recording
the first "full" browser will have the same effect in all cases that
matter.