Don't try and directly invoke git in builds
The build script was trying to populate the auto-generated keystore
with the right user name in the CN= field, by retrieving the user's
name from the git config. With some configurations, this just plain
won't work. So, don't try and be so clever, just use "appauth" as
the name, given it is of no real consequence.