Use an intermediary activity to manage the authorization flow
This change fixes issues #106, #104 and #102.

activity via an Intent with using SINGLE_TOP | CLEAR_TOP, we
can dispose of the browser tab without using the NO_HISTORY flag.
The browser flow will survive the user switching tasks, and
will not appear in the back stack after the authorization flow
completes.

as instance state on the activity. If the activity is completely
killed under memory pressure during the authorization flow,
everything should still work as intended once the application's
activities are restored. This removes the need for
PendingIntentStore, which was using a static map to store this
information previously.

when this is resumed without receiving the redirect from the
authorization flow in the browser. This is a signal that the
user canceled the flow by pressing the back button or closing
the tab. Based on this, we can fire a different PendingIntent
to indicate the flow was canceled.

The changes *should* be backwards compatible. We have also
simplified the configuration of the redirect URI receiver
activity, using a manifest merge attribute set in the build
file.